reading CSV file with Spark and executing a query qith group by.

+++ app/models/helper/Util.scala
@@ -2,7 +2,7 @@ package models.helper
 
 import java.io._
 import java.net.URL
-import java.nio.file.{Files, Paths}
+import java.nio.file.{Files, Path, Paths}
 import java.text.{DateFormat, SimpleDateFormat}
 import java.util.Date
 import java.util.zip.ZipInputStream
@@ -31,25 +31,31 @@ object Util {
   }
 
   def unzip(source: String): Unit = {
+
     try {
-      val zipFile: InputStream = new FileInputStream(new File(source))
-      val zis = new ZipInputStream(zipFile)
+      val dir: Path = new File(source.replace(".zip", "")).toPath
+      if (Files.notExists(dir) && !Files.isDirectory(dir)) {
+        val zipFile: InputStream = new FileInputStream(new File(source))
+        val zis = new ZipInputStream(zipFile)
 
-      Stream.continually(zis.getNextEntry).takeWhile(_ != null).foreach { file =>
-        if (!file.isDirectory) {
-          val outPath = Paths.get("downloads").resolve(file.getName)
-          val outPathParent = outPath.getParent
-          if (!outPathParent.toFile.exists()) {
-            outPathParent.toFile.mkdirs()
+        Stream.continually(zis.getNextEntry).takeWhile(_ != null).foreach { file =>
+          if (!file.isDirectory) {
+            val outPath = Paths.get("downloads").resolve(file.getName)
+            val outPathParent = outPath.getParent
+            if (!outPathParent.toFile.exists()) {
+              outPathParent.toFile.mkdirs()
+            }
+
+            val outFile = outPath.toFile
+            val out = new FileOutputStream(outFile)
+            val buffer = new Array[Byte](4096)
+            Stream.continually(zis.read(buffer)).takeWhile(_ != -1).foreach(out.write(buffer, 0, _))
           }
-
-          val outFile = outPath.toFile
-          val out = new FileOutputStream(outFile)
-          val buffer = new Array[Byte](4096)
-          Stream.continually(zis.read(buffer)).takeWhile(_ != -1).foreach(out.write(buffer, 0, _))
         }
+        println("Decompress complete")
+      } else {
+        println("Dir is already decompressed.")
       }
-      println("Decompress complete")
     } catch {
       case e: Exception => println(s"Error occurred on decompressing the file [${source}]: [${e}]")
     }
